Refactor operations as class attributes
