+++ relentless/simulate/__init__.py
@@ -45,19 +45,7 @@ from .simulate import (Simulation,
                        Barostat,
                        BerendsenBarostat,
                        MTKBarostat)
-
-from . import dilute
-from .dilute import Dilute
-
-from . import hoomd
-from .hoomd import HOOMD
-
-from . import lammps
-from .lammps import LAMMPS
-
-# setup generic (adapter) objects
-from . import generic
-from .generic import (InitializeFromFile,
+from .simulate import (InitializeFromFile,
                       InitializeRandomly,
                       MinimizeEnergy,
                       AddBrownianIntegrator,
@@ -70,6 +58,12 @@ from .generic import (InitializeFromFile,
                       RunUpTo,
                       AddEnsembleAnalyzer,
                       )
-generic.GenericOperation.add_backend(Dilute)
-generic.GenericOperation.add_backend(HOOMD)
-generic.GenericOperation.add_backend(LAMMPS)
+
+from . import dilute
+from .dilute import Dilute
+
+from . import hoomd
+from .hoomd import HOOMD
+
+from . import lammps
+from .lammps import LAMMPS
